+++ packages/xo-web/src/xo-app/vm/tab-advanced.js
@@ -20,6 +20,7 @@ import {
   formatSize,
   getCoresPerSocketPossibilities,
   osFamily,
+  VIRTUALIZATION_MODE_LABEL,
 } from 'utils'
 import {
   cloneVm,
@@ -394,11 +395,7 @@ export default class TabAdvanced extends Component {
               <tbody>
                 <tr>
                   <th>{_('virtualizationMode')}</th>
-                  <td>
-                    {vm.virtualizationMode === 'pv'
-                      ? _('paraVirtualizedMode')
-                      : _('hardwareVirtualizedMode')}
-                  </td>
+                  <td>{_(VIRTUALIZATION_MODE_LABEL[vm.virtualizationMode])}</td>
                 </tr>
                 {vm.virtualizationMode === 'pv' && (
                   <tr>